Edit CSS further, add several secondary functions and events
Made additional updates to the CSS mainly (with minor HTML edits, such as
renaming or removing), and added the structure for several functions for the
script file. Also began working on cutting down global variables that aren't
used more than a couple of times/aren't important enough to be declared.