Add Filtering Functionality
Added the ability to edit books already added to the library by adding three
different dataset attributes to the add-btn HTML tag. This should store the
myLibrary index of the book being edited, the index of the visual book
card within the bookshelf container, and the dataset-id value of the book DIV
that is created when adding a new book to the library initially. This allows
these values to be stored and referenced when updating a book, to update both
the visual book card as well as the properties within the myLibrary array.

Also added filtering functionality (by title, author, or tag), and currently
working on the sorting functionality.