Add Sorting Functionality, Update Functions
Added the ability to sort the library by date added, title, and author,
deleting the visual bookshelf and recreating it each time the sorting is
changed. Also update several functions, including cutting down the function
that updates a book, and added a function to track how many books are
currently in the user's library.