Add Local Storage Capability, Make Minor Edits
Added the ability for the theme, current unique book ID count, and library
array to be saved to local storage whenever any of those is updated/changed.

Also made minor tweaks, such as adding styling for a hover effect for the
edit/delete buttons on books, made the book card height slightly smaller,
and began the process of moving Javascript code around to a more desirable
order.